chore(recursive): refactor to modules
